move esql tests
